Add docker-compose setup
This commit adds a docker-compose.yml file to standardize and simplify the way our services are built and deployed. Also included are necessary configuration files for Docker, Nginx and cron jobs. Dockerfile sets up a Node.js based Docker image with necessary dependencies installed. Nginx is configured to proxy requests to our Node.js service and cronjobs are set to run tasks at specified intervals. The aim is to make the application easier to deploy and manage in a containerized environment.